[symfony] differentiate symfony 3 vs 4 integrations
